Bug: 337098550
Merged-In: Ia2f4731b0893691623c6a72e0009e1d38712feb7

+++ chpp/clients/gnss.c
@@ -400,7 +400,7 @@ static void chppGnssGetCapabilitiesResult(
                 CHPP_GNSS_DEFAULT_CAPABILITIES);
     if (result->capabilities != CHPP_GNSS_DEFAULT_CAPABILITIES) {
       CHPP_LOGE("GNSS capabilities 0x%" PRIx32 " != 0x%" PRIx32,
-                result->capabilities, CHPP_GNSS_DEFAULT_CAPABILITIES);
+                result->capabilities, (uint32_t)CHPP_GNSS_DEFAULT_CAPABILITIES);
     }
 
     clientContext->capabilitiesValid = true;
